Fix lerp_color signed math, buffer overflow, and double syscall
- lerp_color: use int intermediates instead of uint8_t to prevent
  unsigned underflow when a color channel decreases across the ramp
- temp_ramp_color: cast temp_c to int before subtracting TEMP_COLD
  to avoid uint8_t wrap when temperature is below 30C
- format_rate buffer: bump from 6 to 8 bytes to fit "999.9K\0"
- collect_data: store get_apt_update_count() result in a local to
  avoid parsing the apt status file twice
- Remove unused ROW_SPARK_BOT define

+++ project/theme.c
@@ -102,11 +102,11 @@ uint16_t threshold_color(uint32_t value, uint32_t warn_th, uint32_t crit_th) {
 }
 
 uint16_t lerp_color(uint16_t a, uint16_t b, float t) {
-    uint8_t ar = (a >> 11) & 0x1F, ag = (a >> 5) & 0x3F, ab = a & 0x1F;
-    uint8_t br = (b >> 11) & 0x1F, bg = (b >> 5) & 0x3F, bb = b & 0x1F;
-    uint8_t r = ar + (int)((br - ar) * t);
-    uint8_t g = ag + (int)((bg - ag) * t);
-    uint8_t bv = ab + (int)((bb - ab) * t);
+    int ar = (a >> 11) & 0x1F, ag = (a >> 5) & 0x3F, ab = a & 0x1F;
+    int br = (b >> 11) & 0x1F, bg = (b >> 5) & 0x3F, bb = b & 0x1F;
+    int r = ar + (int)((br - ar) * t);
+    int g = ag + (int)((bg - ag) * t);
+    int bv = ab + (int)((bb - ab) * t);
     return (r << 11) | (g << 5) | bv;
 }
 
@@ -115,7 +115,7 @@ static float clampf(float v, float lo, float hi) { return v < lo ? lo : v > hi ?
 uint16_t temp_ramp_color(uint8_t temp_c) {
     if (temp_c <= TEMP_COOL)
         return lerp_color(theme.tempRamp[0], theme.tempRamp[1],
-                          clampf((temp_c - TEMP_COLD) / (float)(TEMP_COOL - TEMP_COLD), 0.0f, 1.0f));
+                          clampf(((int)temp_c - TEMP_COLD) / (float)(TEMP_COOL - TEMP_COLD), 0.0f, 1.0f));
     if (temp_c <= TEMP_WARM)
         return lerp_color(theme.tempRamp[1], theme.tempRamp[2], (temp_c - TEMP_COOL) / (float)(TEMP_WARM - TEMP_COOL));
     if (temp_c <= TEMP_HOT)
